[added] More server customization options. [added] Mini-suit. [added] Broken skill mods.
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1507 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/feature/FeatureManager.h
@@ -81,8 +81,8 @@ public:
 		return floatFeature->contains(key);
 	}
 
-	inline int getFloatFeature(const String& key) {
-		return (int)floatFeature->get(key);
+	inline float getFloatFeature(const String& key) {
+		return floatFeature->get(key);
 	}
 
 	inline bool hasSetFeature(const String& key, const String& feature) {
